consolidated url_replace to hawc, moved add_class to bs4
