+++ hawc/apps/animal/templates/animal/endpoint_list_v2.html
@@ -1,7 +1,5 @@
 {% extends 'assessment-rooted.html' %}
 
-{% load url_replace %}
-
 {% block content %}
 <h2>Assessment endpoints v2</h2>
 <div id="main"></div>
