+++ hawc/apps/assessment/templates/assessment/assessment_log_list.html
@@ -1,6 +1,5 @@
 {% extends 'assessment-rooted.html' %}
 
-{% load url_replace %}
 {% load crispy_forms_tags %}
 
 {% block content %}
