+++ hawc/apps/common/templatetags/bs4.py
@@ -1,6 +1,7 @@
 """
 Twitter Bootstrap 4 - helper methods
 """
+import re
 from textwrap import dedent
 from uuid import uuid4
 
@@ -102,3 +103,24 @@ def plotly(fig: Figure | None, **kw) -> str:
     <script>document.addEventListener("{event}", {func}, false);</script>"""
         )
     )
+
+
+class_re = re.compile(r'(?<=class=["\'])(.*)(?=["\'])')
+
+
+@register.filter
+def add_class(value, css_class):
+    """http://djangosnippets.org/snippets/2253/
+    Example call: {{field.name|add_class:"col-md-4"}}"""
+    string = str(value)
+    match = class_re.search(string)
+    if match:
+        m = re.search(
+            rf"^{css_class}$|^{css_class}\s|\s{css_class}\s|\s{css_class}$",
+            match.group(1),
+        )
+        if not m:
+            return mark_safe(class_re.sub(match.group(1) + " " + css_class, string))
+    else:
+        return mark_safe(string.replace(">", f' class="{css_class}">'))
+    return value
