+++ hawc/templates/includes/paginator.html
@@ -1,5 +1,3 @@
-{% load url_replace %}
-
 {% if is_paginated %}
 <ul class="pagination justify-content-center mb-0">
